renamed
I am adding them outside of the folder for them so it doesn't update the
files. The only issue with them is that when i open the program on my
local machine they don't show up as updated names, but they are updated
and the names are inside.  I will try to figure out why it will not
update online if the files are updated on Thursday.